requirements.txt
streamlit as st
pandas as pd
numpy as np
matplotlib.pyplot as plt